server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const { mapDBToModel } = require('../../utils');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this.cacheService = cacheService;
   }
 
   async verifyNoteAccess(noteId, userId) {
@@ -59,20 +60,27 @@ class NotesService {
     if (!result.rows[0].id) {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
+    await this.cacheService.delete(`notes:${owner}`);
     return result.rows[0].id;
   }
 
   async getNote(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-      LEFT JOIN collaborations ON collaborations.note_id = notes.id
-      WHERE notes.owner = $1 OR collaborations.user_id = $1
-      GROUP BY notes.id`,
-
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
+    try {
+      const result = await this.cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+        LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner = $1 OR collaborations.user_id = $1
+        GROUP BY notes.id`,
+        values: [owner],
+      };
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+      await this.cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -93,7 +101,7 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updateAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, update_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, update_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updateAt, id],
     };
     const result = await this._pool.query(query);
@@ -101,18 +109,22 @@ class NotesService {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
     const res = await this.getNoteById(id);
+    const { owner } = result.rows[0];
+    await this.cacheService.delete(`notes:${owner}`);
     return res;
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
     const result = await this._pool.query(query);
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+    const { owner } = result.rows[0];
+    await this.cacheService.delete(`notes:${owner}`);
   }
 }
